Add `RegisterSubjectForm` view - On the `SubjectController` `delete` and `create` functions to check if the subject is associated with teachers or batch and if the subject is deleted on soft delete not to create - Get deleted subjects from the `index` function to render on the subjects `Index` page - Add `restore` function to restore soft deleted subjects and add their routes - Add `subjects.js` store file - Remove subjects `Add` page - Update subjects `Update` page

+++ app/Http/Controllers/SubjectController.php
@@ -12,31 +12,43 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
 use Inertia\Inertia;
+use Inertia\Response;
 
 class SubjectController extends Controller
 {
-    public function index(Request $request)
+    public function index(Request $request): Response
     {
         // Get search key
         $searchKey = $request->input('search');
 
         // Get subjects
-        $subjects = Subject::select('id', 'full_name', 'short_name')->where('full_name', 'like', '%'.$searchKey.'%')->paginate(10);
+        $subjects = Subject::select('id', 'full_name', 'short_name', 'category', 'tags')->where('full_name', 'like', '%'.$searchKey.'%')->paginate(10);
+
+        // Get deleted subjects with search key
+        $deletedSubjects = Subject::onlyTrashed()->select('id', 'full_name', 'short_name')->where('full_name', 'like', '%'.$searchKey.'%')->paginate(10);
 
         return Inertia::render('Subject/Index', [
             'subjects' => $subjects,
+            'deletedSubjects' => $deletedSubjects,
         ]);
     }
 
     public function create(CreateRequest $request): RedirectResponse
     {
         try {
-            Subject::create([
-                'full_name' => $request->full_name,
-                'short_name' => $request->short_name,
-                'category' => $request->category,
-                'tags' => $request->tags,
-            ]);
+            // Check if subject is deleted
+            $subject = Subject::withTrashed()->where('full_name', $request->full_name)->first();
+
+            if (! $subject) {
+                Subject::create([
+                    'full_name' => $request->full_name,
+                    'short_name' => $request->short_name,
+                    'category' => $request->category,
+                    'tags' => $request->tags,
+                ]);
+            } else {
+                return redirect()->back()->with('error', 'Subject already exists.');
+            }
 
             return redirect()->back()->with('success', $request->full_name.' added successfully');
         } catch (Exception $e) {
@@ -99,10 +111,38 @@ class SubjectController extends Controller
                 return redirect()->back()->with('error', 'Subject not found.');
             }
 
-            // Delete subject
-            $subject->delete();
+            // Check if subject is associated with teacher or batch
+            if ($subject->teachers || $subject->batch) {
+                // Delete subject with soft delete
+                $subject->delete();
+            }
+
+            // Delete subject permanently
+            $subject->forceDelete();
 
             return redirect()->back()->with('success', 'Subject deleted successfully');
+        } catch (Exception $e) {
+            Log::error($e->getMessage());
+
+            return redirect()->back()->with('error', 'Something went wrong. Please try again.');
+        }
+    }
+
+    // Restore subject
+    public function restore($id): RedirectResponse
+    {
+        try {
+            // Check if subject exists
+            $subject = Subject::withTrashed()->find($id);
+
+            if (! $subject) {
+                return redirect()->back()->with('error', 'Subject not found.');
+            }
+
+            // Restore subject
+            $subject->restore();
+
+            return redirect()->back()->with('success', 'Subject restored successfully');
         } catch (Exception $e) {
             Log::error($e->getMessage());
